Implemented pearson collaborative filtering algorithm and example using the MovieLens dataset
The results were off for awhile and are still a bit suspect, but, then again,
this could just be weaknesses in the algorithm.  Coming up with a way
of grading the algorithms would be most excellent.  I guess that will be
an area of research after implementing a few additional filter algos.